Capture metalsmith python logging
The metalsmith_instances module calls into metalsmith and there is
currently no logging mechanism to see what is happening. This change
is an attempt to improve this by configuring logging and adding the
output to the result dict as 'logging'.

This change is identical to tripleo-ansible change
https://review.opendev.org/#/c/740606/ so that there is no regression
when tripleo-ansible deletes its copy.

Change-Id: I8694810554b2e09de31786f7e1dd1cbbfa0e8f60
Related-Bug: #1879472