Add a Centos8 UEFI based job
Turns out we seem to be using bios boot with all of the metalsmith
jobs, but we need to further test UEFI and ensure that we cover
uefi and partition/wholedisk code in the ramdisk... and these
jobs are actually not bad for this purpose.

Since we ultimately want to replace the centos7 job, but centos8.2
has some incompatabilities with our IPA image and bootloader
installation process with partition images, we need to resolve
those issues before making this job voting OR replacing the
centos7 job.

Change-Id: Ibe6c543179c811c3ef43feca173a4e744c7f5671