Change metalsmith nodeset to Focal
By default, metalsmith jobs relied upon the devstack
default which would pull in bionic, but as we want to
test UEFI soon, we need to have focal jobs for the working
uefi driver PXE network stack.

Also disables dstat as we don't presently use it and it seems
broken on focal. :\

Change-Id: I411e2bc9c15ee4fb77fba980e1924823fe16ba3f