Set binding:host_id on VIFs
Set the 'binding:host_id' on neutron port VIFs when
they are created, or update it for existing ports.

This ensures IP allocation is not deferred by neutron.

If IP allocation is deferred, the config-drive network
metadata will not include enough information to properly
bootstrap the instance networking.

Story: 2009715
Task: 44089
Closes-Bug: #1952573
Change-Id: I7411c1f2fbc57dad0504fe61e278308030f059bf