CI: Extend metalsmith CI job timeout
Metalsmith has always been a fairly long running job, in part
because it hast to build images, and then deploy them.

It appears, we are seeing increased occurances of the job
timing out due to what appears to be just slightly slower IO
performance. Overall major steps are still executing in an
acceptable amount of time, just towards the further end of
what we typically expect to see in CI.

As such, extend the overall job timeout a little bit to give
it time to hopefully finish. Extends the job timeout 1000 seconds,
which is a bit over sixteen more minutes to allow the job to
complete cleaning.

Change-Id: I5c7928e341ffcbc94f1beb8a33a8d10d89bae8a7