stable-only: Constrain the upper Ansible version
On CI nodes, we get the latest ansible version, generally,
such as 2.9.15.

However, this code was originally developed on 2.8.x, *and*
utilizes an older openstacksdk version which comes from our
requirements and stable branch upper constraints, for example
openstacksdk 0.101.0 being the current version on the Zed branch.

However, Ansible 2.9 is incompatible with versions of the SDK
gerater than 0.99.0. To make matters worse, Ansible expects
SDK version greater than 1.0 because there are breaking changes
for the SDK which they navigated at the 1.0 route.

So since we're on an older stable branch, lets constrain our
test run ansible version so we don't end up failing the CI
job because our base ansible is too new for the software we
are trying to test.

Presently planning to backport this on stable branches so we
can unbrick IPA stable branches.

Change-Id: I67a3397f53e03270ac3ce540042f188991cd2bac