CI: Ask ironic devstack to set node owner
The metalsmith CI job yaml documents contain a default OS_CLOUD,
which is fine, until Ironic sets the new Secure RBAC model to
enforcing by default. At which time, metalsmith jobs are not able
to see or modify nodes they previously could running as a project
scoped admin.

Executing the job as a project scoped admin is fine. The challenge
is the user's project needs to be granted access to see and use
the baremetal ndoe.

And so, to do this, we'll utilize a flag being added to ironic's
devstack script to trigger the node to be configured with an owner
which will allow metalsmith's requests to be serviced (in that the
baremetal node *will* be visible, and can be scheduled upon).

Change-Id: I5248894c824381f4a283b2746102df04962f20b0