Use bootloader element instead of grub2
bootloader is recommended since it aims to build a valid grub config
instead of just installing the packages.

Also, mark legacy job as non-voting for the time being

The legacy job is broken because of a know issue with tinycore related
to the glibc library.  Moving it to non-voting for the time being while
we evaluate if it's worth trying to fix it or jsut remove it.

Also, install diskimage-builder from the checked-out repo, so that
depends on works as expected.

Change-Id: I642a1579e95dfff2188206313ea6d315b0533a5f