Error with clarity when a bad upgrade was encountered
Metalsmith is generally used as a frontend for baremetal
provisioning such that a user can ask for baremetal to be
deployed with a simplified command line. The challenge is
it artificially hydrates together data from several queries
to return a list of instances which have been deployed
or are managed using metalsmith. The challenge is if someone
didn't create an allocation record, or if someone used the
metalsmith modules to do so when the node was in a maintenance
state, for example if BMC communication was not working.

The underlying challenge is this results in a lack of an allocation
record from being created. This can be manually corrected, and instead
of providing a vague error, metalsmith now provides a detailed error
with the command to remedy the error state through manual creation of
the allocation record, which allows "metalsmith list" to work as
expected.

Change-Id: I033a9925fe798f329b7bb9a629da0cd0f1202683