Fix metadata generation
Metalsmith has long using an incorrect pattern to generate metadata.

And honestly its sort of Ironic's fault. Ironic doesn't do it for the
user, which is horrible UX.

Anyhow, the path metalsmith took was to populate everything into
neutron, then issue a vif binding request to ironic. This process of
populating eveyrthing into neutron also triggered the port address to
be assigned through providing a host:binding_id value which begins the
process to bind a port. And then metalsmith would query the ports and
data together to generate a functional network_metadata payload.

Where this goes sideways is while attempting to fix bug #2106073 as it
pertains to Ironic, any fix ultimately breaks metalsmith's ability
to take its prior path and to be able to directly use it for metadata
generation. Ultimately, the path metalsmith needs to take is similar
to nova's process, which is to cross-reference the VIF on the port
and use that to identify the MAC address.

This change does that, in a minimal/no-frills fashion.

Closes-Bug: #2107483
Change-Id: I2e51d8d1920dbf576cb625720682c078cef07357